Sunday 8
Create VisitorFine, Booking

+++ src/main/java/com/api/hotel/repo/BookingRepo.java
@@ -3,6 +3,6 @@ package com.api.hotel.repo;
 import com.api.hotel.model.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepo extends JpaRepository<Booking,Long> {
+public interface BookingRepo extends JpaRepository<Booking, Integer> {
 
 }
